Fold atomic creation into issue tools

Copilot-Session: f71d9868-eef8-4fb0-84c6-df7c9a6a0ade
